[6674] * Replace shell/make scripts use for svn_version.h generation by use src\tools\gensvnrevision (used at Windows). Note: --enable-svn-extract configure option removed, src\tools\gensvnrevision proper work in all cases without this option support. * Fixed: some GCC warnings.

+++ src/shared/ByteBuffer.h
@@ -294,8 +294,8 @@ class ByteBuffer
             {
                 if(guid & 0xFF)
                 {
-                    _storage[mask_position] |= (1<<i);
-                    *this << ((uint8)(guid & 0xFF));
+                    _storage[mask_position] |= uint8(1<<i);
+                    *this << uint8(guid & 0xFF);
                 }
 
                 guid >>= 8;
